Added some files common to all projects/branches.
description of improvements and collection of results (greek).xlsx is an excel
sheet containing the results of running the different versions and a description
of their differences (in Greek).

exercise description (greek).pdf contains the description of two exercises (in
Greek), the second of which is the 'voltes' exercise that gave me the idea of
creating this set of projects.

repository contents.txt is a small description of this repository. (or rather of
how this repository will be)