Added the old Visual Studio 2008 solution I used to use to build the project.
This is for historical reasons. The paths to the source files are obsolete.

Originally I had created a new VS2008 solution for each version and had saved
each revision in a separate folder as, at the time, I was unaware of the
existence of revision control systems.
Most solutions were the identical (except for the unique IDs) and a few had a
flag or two or a spelling error the others didn't. I chose the most common
version and renamed it to 'voltes'.

The original names for the solutions/projects will be used as the branch names.